Player cannot change the cycle direction before the game begins. LinePath now holds a second path that holds the coordinates for where on the canvas the path will be drawn. This was done to avoid recomputation related to converting coordinates from the Grid to the screen. LinePath now draws its path. Added helper static  methods to Point. Replaced GestureDetector with MultiSwipeDetector. This class allows multiple fingers to be tracked for swipes. The Game now allows for two cycles to be played. player zero swiped the top half of teh screen and player 2 swiped teh bottom half of the screen. The min sdk is now 8 in order to support newer motion event functions.

+++ app/src/main/java/com/yckir/cyclebattledemo/PathVertex.java
@@ -35,10 +35,10 @@ public class PathVertex {
         Time=time;
         Direction=direction;
         mWidth=DEFAULT_WIDTH;
-        mDrawingLeft=Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, X - mWidth);
-        mDrawingTop=Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, Y - mWidth);
-        mDrawingRight=Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, X + mWidth);
-        mDrawingBottom=Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, Y + mWidth);
+        mDrawingLeft=(int)Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, X - mWidth);
+        mDrawingTop=(int)Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, Y - mWidth);
+        mDrawingRight=(int)Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, X + mWidth);
+        mDrawingBottom=(int)Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, Y + mWidth);
     }
 
 
@@ -91,8 +91,8 @@ public class PathVertex {
      */
     public void setX(double x) {
         X = x;
-        mDrawingLeft=Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, X - mWidth);
-        mDrawingRight=Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, X + mWidth);
+        mDrawingLeft=(int)Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, X - mWidth);
+        mDrawingRight=(int)Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, X + mWidth);
     }
 
 
@@ -101,8 +101,8 @@ public class PathVertex {
      */
     public void setY(double y) {
         Y = y;
-        mDrawingTop=Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, Y - mWidth);
-        mDrawingBottom=Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, Y + mWidth);
+        mDrawingTop=(int)Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, Y - mWidth);
+        mDrawingBottom=(int)Tile.convert(Grid.GAME_GRID_TILE, GameFrame.SCREEN_GRID_TILE, Y + mWidth);
     }
